Add workspace support for the tracking store
If workspaces are not enabled, additional filters are skipped where
reasonable. When workspaces are enabled, there is always a filter on the
workspace column for the resource itself or its parent (e.g. a run's
experiment).

Additionally, artifact isolation when MLflow is serving artifacts must
be scoped under /workspaces/<workspace name> to ensure isolation. When
MLflow is not serving artifacts, then that is the default behavior with
the default artifact root but workspace providers can override this.

Additionally, a Flask before request handler and a FastAPI middleware
set the workspace context variable based on the X-MLFLOW-WORKSPACE
header value.

The default workspace is "default" for the SqlAlchemyStore
implementation of the workspace provider, but it's possible that the
workspace provider does not support defaults (raises
NotImplementedError). In that case, the workspace is set to `None` and
the tracking store is responsible for erroring when a workspace is
needed and one is not provided.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>